requirements.txt
Django==1.8
djangorestframework==3.5.3
wheel==0.24.0
